requirements.txt
PySide6
qtawesome
werkzeug
matplotlib
python-dotenv